Cycles Denoising: Hopefully fix Multi-GPU denoising
The problem there is that currently tiles get allocated on the GPU that's used to render them.
However, if a GPU is supposed to denoise a tile, it needs all 8 neighbors in its memory as well.

Therefore, the code now allocates and copies the tiles on the denoising GPU as well.