Cycles: Extend Persistent Images to other data
When changing frame while rendering in the Viewport, Cycles is often able to keep most objects in their current state and only update a few of them.
This option enables similar behaviour for F12 rendering, which can reduce rendertimes a lot in certain cases.

Note that this is fairly experimental and might still be buggy.