Cycles: Add dithered scrambling option for the Sobol sampler
Since there is only one Sobol sequence, we need to scramble it somehow to get different sequences for each pixel in order to avoid correlation artifacts.
By default, each dimension is shifted by a per-pixel random value.

However, by picking shifts that are as different as possible for neighboring pixels, we can change the noise pattern to contain less low-frequency noise, which tends to be more noticeable.

Note that this is still experimental, I'm fairly sure that it should work better.