Cycles: Add option to specify scrambling distance
As explained in the dithering commit, each sample is shifted to decorrelate pixels. This commit makes it possible to reduce the amount
of shifting, which makes neighboring paths more similar, which can improve performance especially on GPUs. Also, lower values can
reduce noise.

However, there are also downsides - if you pick the value too low, you'll get noticeable artifacts, commonly in the form of lighting/shadows
suddenly popping up over the course of rendering, which makes it hard to tell when the render is done.

Therefore, you should be careful with this setting - I wouldn't recommend ever going below 0.25, and staying at 1 if you're doing 64 samples or less.