Cycles: Add option to generate AOV passes during rendering
With this patch, you can generate arbitraty output passes in addition to the default Cycles passes.
To do so, add your AOV pass in the Render Layer tab, specify the type (Color or Value) and the name.

Then, for each shader that should contribute to the pass, add an AOV Output Node, specify the Name of
the pass and connect the color or value that you want to output. The pass will then be part of the
render as "AOV <Name>".

Note that only colors and values are supported, there is NO support for connecting BSDFs to the output.