Cycles: Allow to connect a tangent input to the Normal Map node
This is just a small usability hack: When using a Normal Map node inside a Node Group,
you can't change the UV Map that's being used for one specific instance of the Node Group,
which means you need to duplicate the node group (and lose the advantage of having groups at all) for each UV map.

With this patch, you can have a Tangent node on the outside of the Node Group, specify the UV map there and then
use the tangent inside the group.

However, there is one important point to keep in mind - this ONLY works if there is no other node between the Tangent
and Normal Map. The reason for that is that the Normal Map needs not just the tangent itself, but also other info.
Therefore, this functionality is implemented in the Shader optimizer, which merges the Tangent Node (and its UV setting)
into the Normal Map node, but only if there is a direct connection.