Cycles: Allow to use host memory for CUDA rendering
This commit allows to fall back to host memory (aka RAM) when the GPU memory (aka vRAM) is full. Performance is affected, but at least it's possible to render larger scenes at all.

To use it, specify the percentage of host memory that may be used. Note that going too high might cause system instability.

Credit for this patch goes to Stefan Werner.